+++ CHANGELOG.md
@@ -23,6 +23,9 @@
 
 ### Fixed
 
+* Forward worker errors to test output in the test runner.
+  [#XXXX](https://github.com/wasm-bindgen/wasm-bindgen/pull/XXXX)
+
 ### Removed
 
 ## [0.2.106](https://github.com/wasm-bindgen/wasm-bindgen/compare/0.2.105...0.2.106)